Split large controller out into per-endpoint files
